requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
pygments
django_extensions
ipython
pillow
django-filter
celery
redis
django-celery-beat
django-admin-multi-select-filter
